Revert "Latest version with virtual connection"
This reverts commit 5920598302cd563c705742b6533ceb1ceff1e14b.

+++ src/edu/upc/epsevg/prop/hex/UnitTesting.java
@@ -39,7 +39,7 @@ public class UnitTesting {
 
         HexGameStatus gs = new HexGameStatus(board, PlayerType.PLAYER1);
 
-        double h = PlayerID.dijkstraDistanceToWin(gs, -1);//PlayerMinimax.dijkstraDistanceToWin(gs, -1);
+        int h = PlayerID.dijkstraDistanceToWin(gs, -1);//PlayerMinimax.dijkstraDistanceToWin(gs, -1);
         System.out.println(h);
     }
